requirements.txt
jupyter==1.0.0
numpy==1.16.1
Pillow==5.4.1
tensorflow==1.12.0
jinja2==2.10

